+++ modules/pytket-projectq/tests/projectq_backend_test.py
@@ -52,10 +52,10 @@ def circuit_gen(measure: bool = False) -> Circuit:
 def test_statevector() -> None:
     c = circuit_gen()
     b = ProjectQBackend()
-    state = b.get_state(c)
+    state = b.run_circuit(c).get_state()
     assert np.allclose(state, [math.sqrt(0.5), 0, 0, math.sqrt(0.5)], atol=1e-10)
     c.add_phase(0.5)
-    state1 = b.get_state(c)
+    state1 = b.run_circuit(c).get_state()
     assert np.allclose(state1, state * 1j, atol=1e-10)
 
 
@@ -106,11 +106,14 @@ def test_ilo() -> None:
     b = ProjectQBackend()
     c = Circuit(2)
     c.X(1)
+    res = b.run_circuit(c)
     assert np.allclose(
-        b.get_state(c), np.asarray([0, 1, 0, 0], dtype=complex), atol=1e-10
+        res.get_state(),
+        np.asarray([0, 1, 0, 0], dtype=complex),
+        atol=1e-10,
     )
     assert np.allclose(
-        b.get_state(c, basis=BasisOrder.dlo),
+        res.get_state(basis=BasisOrder.dlo),
         np.asarray([0, 0, 1, 0], dtype=complex),
         atol=1e-10,
     )
@@ -126,8 +129,9 @@ def test_swaps_basisorder() -> None:
     CliffordSimp(True).apply(c)
     assert c.n_gates_of_type(OpType.CX) == 1
     c = b.get_compiled_circuit(c)
-    s_ilo = b.get_state(c, basis=BasisOrder.ilo)
-    s_dlo = b.get_state(c, basis=BasisOrder.dlo)
+    res = b.run_circuit(c)
+    s_ilo = res.get_state(basis=BasisOrder.ilo)
+    s_dlo = res.get_state(basis=BasisOrder.dlo)
     correct_ilo = np.zeros((16,))
     correct_ilo[4] = 1.0
     assert np.allclose(s_ilo, correct_ilo)
@@ -184,11 +188,6 @@ def test_shots_bits_edgecases(n_shots, n_bits) -> None:
     assert np.array_equal(res.get_shots(), correct_shots)
     assert res.get_shots().shape == correct_shape
     assert res.get_counts() == correct_counts
-
-    # Direct
-    assert np.array_equal(projectq_backend.get_shots(c, n_shots), correct_shots)
-    assert projectq_backend.get_shots(c, n_shots).shape == correct_shape
-    assert projectq_backend.get_counts(c, n_shots) == correct_counts
 
 
 def test_backend_info() -> None: